+++ packages/components/src/sandbox/index.native.js
@@ -14,6 +14,7 @@ import {
 	useState,
 	useEffect,
 } from '@wordpress/element';
+import { usePreferredColorScheme } from '@wordpress/compose';
 
 /**
  * Internal dependencies
@@ -162,6 +163,7 @@ function Sandbox( {
 	type,
 	url,
 } ) {
+	const colorScheme = usePreferredColorScheme();
 	const ref = useRef();
 	const [ height, setHeight ] = useState( 0 );
 	const [ contentHtml, setContentHtml ] = useState( getHtmlDoc() );
@@ -171,10 +173,14 @@ function Sandbox( {
 		windowSize.width >= windowSize.height
 	);
 	const wasLandscape = useRef( isLandscape );
-	// On Android, we need to recreate the WebView when the device rotates, otherwise it disappears.
-	// For this purpose, the key value used in the WebView will change when the device orientation gets updated.
+	// On Android, we need to recreate the WebView on any of the following actions, otherwise it disappears:
+	// - Device rotation
+	// - Light/dark mode changes
+	// For this purpose, the key prop used in the WebView will be updated with the value of the actions.
 	const key = Platform.select( {
-		android: `${ url }-${ isLandscape ? 'landscape' : 'portrait' }`,
+		android: `${ url }-${
+			isLandscape ? 'landscape' : 'portrait'
+		}-${ colorScheme }`,
 		ios: url,
 	} );