+++ packages/react-native-aztec/ios/RNTAztecView/RCTAztecViewManager.m
@@ -27,6 +27,7 @@ RCT_EXPORT_VIEW_PROPERTY(maxWidth, CGFloat)
 RCT_EXPORT_VIEW_PROPERTY(fontFamily, NSString)
 RCT_EXPORT_VIEW_PROPERTY(fontSize, CGFloat)
 RCT_EXPORT_VIEW_PROPERTY(fontWeight, NSString)
+RCT_EXPORT_VIEW_PROPERTY(lineHeight, CGFloat)
 
 RCT_EXPORT_VIEW_PROPERTY(disableEditingMenu, BOOL)
 RCT_REMAP_VIEW_PROPERTY(textAlign, textAlignment, NSTextAlignment)